Added delete friend functionality Added is friend with query in typeDefs, resolver and queries
Added a function in the user model to check if the friendId is one of the user friends ids

Updated the user interface in use.jsx to display the delete user button when the users are friends. If not, the interface will show the add user button for all users except the currently logged-in user.